int overflow fixes (#608)

+++ ethereum/statetransition/src/main/java/tech/pegasys/artemis/statetransition/util/SlotProcessorUtil.java
@@ -30,7 +30,8 @@ public final class SlotProcessorUtil {
       List<Bytes32> latest_block_roots = state.getLatest_block_roots();
 
       latest_block_roots.set(
-          ((int) state.getSlot() - 1) % Constants.LATEST_BLOCK_ROOTS_LENGTH, previous_block_root);
+          Math.toIntExact((state.getSlot() - 1) % Constants.LATEST_BLOCK_ROOTS_LENGTH),
+          previous_block_root);
       state.setLatest_block_roots(latest_block_roots);
     }
